requirements.txt
# ADS Framework Requirements - CPU-Optimized for Linux Laptops
# Install with: pip install -r requirements.txt

# ==================== CORE DEPENDENCIES ====================
# PyTorch CPU-only build (specified separately due to size)
# Install with: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# ==================== LANGUAGE MODELS & TRANSFORMERS ====================
transformers>=4.36.0        # HuggingFace transformers for LLMs
datasets>=2.14.0            # HuggingFace datasets library
tokenizers>=0.14.0          # Fast tokenization
sentence-transformers>=2.2.0 # Embeddings (optional, for dense retrieval)
peft>=0.7.0                 # Parameter-efficient fine-tuning

# ==================== DATA & SCIENCE ====================
numpy>=1.24.0               # Numerical computing
scipy>=1.11.0               # Scientific computing
scikit-learn>=1.3.0         # ML utilities
pandas>=2.0.0               # Data manipulation
rank-bm25>=0.2.2            # BM25 sparse retrieval

# ==================== TRAINING & OPTIMIZATION ====================
torch-optimization>=1.0.0   # Advanced optimizers (optional)
tqdm>=4.66.0                # Progress bars

# ==================== UTILITIES ====================
accelerate>=0.24.0          # Distributed training utilities
safetensors>=0.4.0          # Efficient model serialization

# ==================== LOGGING & MONITORING ====================
tensorboard>=2.14.0         # TensorBoard integration
wandb>=0.15.0               # Weights & Biases (optional)

# ==================== MISCELLANEOUS ====================
pyyaml>=6.0                 # YAML config files
requests>=2.31.0            # HTTP requests
python-dotenv>=1.0.0        # Environment variables
tqdm>=4.66.0                # Progress visualization

# ==================== DEVELOPMENT (Optional) ====================
# pytest>=7.4.0              # Testing framework
# black>=23.0.0              # Code formatting
# flake8>=6.0.0              # Linting
# mypy>=1.5.0                # Type checking
